fixed spread method
Before this change a node did not spread a message only if one of the nodes calculated was the requester. This is wrong: for example, a node which calculates the two subsequent nodes after the requester would send a message to them, but they already received it from the requester.