renamed and moved a few things
removed final keywords where useless. A converseListener (please change this name because its not clear what it means) for me is a listener set by the user listening for valueFound, joinProposal and valueNotFound
SMSNetworkManagementListener is used internally by the network manager in response to node founds and to join agreed. Maybe we can add more things here. And we also need an implementation of this interface (or just define a class).
Moved processRequest and processReply to the listener, since CommandMapper should merely map commands and send them